+++ src/components/bottomSheet/BottomSheet.styles.ts
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import theme from '../../styles/theme';
-import { slideUpAnimation } from '../../utils/slideUpAnimation';
+import { slideUpAnimation } from '../../utils/slideAnimation';
 
 export const Overlay = styled.div`
   width: 100vw;
