+++ src/components/option/index.styles.ts
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+export const Wrapper = styled.div`
+  width: 100%;
+
+  display: flex;
+  flex-wrap: wrap;
+  row-gap: 5px;
+  column-gap: 6px;
+`;
