+++ src/components/option/participantsOption/Pariticipant.tsx
@@ -1,5 +1,5 @@
-import { ParticipantWrapper } from './SelectParticipants.styles';
-import { ParticipantBlockTypes } from './resultOption.types';
+import { ParticipantWrapper } from './participantsOption.styles';
+import { ParticipantBlockTypes } from './participantsOption.types';
 
 const Participant = ({ id, isSelected, onClick }: ParticipantBlockTypes) => {
   return (
