+++ src/components/option/sortOption/SortTimes.styles.ts
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import theme from '../../styles/theme';
+import theme from '../../../styles/theme';
 
 export const Wrapper = styled.div`
   width: 100%;
