+++ src/utils/slideAnimation.ts
@@ -10,3 +10,14 @@ export const slideUpAnimation = keyframes`
         transform: translateY(0);
     }
 `;
+
+export const slideRightAnimation = keyframes`
+    from {
+        opacity: 0.5;
+        transform: translateX(-100%);
+    }
+    to {
+        opacity: 1;
+        transform: translateX(0);
+    }
+`;
